modificacion del api rest

+++ NVC-APIREST/src/main/java/idat/edu/pe/repositorio/CargoRepositorio.java
@@ -3,7 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import idat.edu.pe.modelo.Cargo;
 
-
-public interface CargoRepositorio extends JpaRepository<Cargo,Long>{
+public interface CargoRepositorio extends JpaRepository<Cargo, Integer>{
 
 }
