+++ NVC-APIREST/src/main/java/idat/edu/pe/repositorio/EmpleadosRepositorio.java
@@ -1,7 +1,11 @@
 package idat.edu.pe.repositorio;
+
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import idat.edu.pe.modelo.Empleados;
-public interface EmpleadosRepositorio extends JpaRepository<Empleados,Long> {
+
+public interface EmpleadosRepositorio extends JpaRepository<Empleados, Integer>{
 
 }
